workshop3/requirements.txt
langchain
langchain-community
langchain-core
langchain-aws
boto3
python-dotenv
fastapi
uvicorn[standard]
pydantic
deepeval
pytest
pytest-asyncio
